requirements.txt
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
asyncio==4.0.0
asyncpg==0.30.0
bcrypt==4.3.0
bidict==0.23.1
cached-property==2.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
defusedxml==0.7.1
dnspython==2.7.0
exchangelib==5.5.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
isodate==0.7.2
lxml==6.0.0
oauthlib==3.3.1
passlib==1.7.4
psycopg2==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyspnego==0.11.2
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
redis==4.5.1
requests==2.32.4
requests-oauthlib==2.0.0
requests_ntlm==1.3.0
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
wsproto==1.2.0
